Drop unnecessary `Page.total_count` function
This function doesn’t belong on the `Page` module as it counts the full
result set for the query without applying any pagination. Furthermore,
we’ve decided to remove it as there doesn’t seem to be any great value
in having it here. The user of Chunkr is already passing in the query to
be paginated; rather than call `:count` on that query, we can leave that
as a simple exercise for the user.